models/Lipsync/requirements.txt
fastapi
uvicorn
tqdm
requests
dotenv
syncsdk    # if sync.so SDK is pip installable
